+++ cli/tests/test_run_command.rs
@@ -1,0 +1,180 @@
+// Copyright 2024 The Jujutsu Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+use crate::common::TestEnvironment;
+use crate::common::TestWorkDir;
+
+#[test]
+fn test_run_simple() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+    let fake_formatter = assert_cmd::cargo::cargo_bin("fake-formatter");
+    assert!(fake_formatter.is_file());
+    let fake_formatter_path = fake_formatter.to_string_lossy();
+    work_dir.write_file("A.txt", "A");
+    work_dir.run_jj(&["commit", "-m", "A"]).success();
+    work_dir.write_file("b.txt", "b");
+    work_dir.run_jj(&["commit", "-m ", "B"]).success();
+    work_dir.write_file("c.txt", "test to replace");
+    work_dir.run_jj(&["commit", "-m", "C"]).success();
+    insta::assert_snapshot!(get_log_output(&work_dir), @r"
+    @  zsuskulnrvyrovkzqrwmxqlsskqntxvp
+    ○  kkmpptxzrspxrzommnulwmwkkqwworplC
+    │
+    ○  rlvkpnrzqnoowoytxnquwvuryrwnrmlp
+    │
+    ○  qpvuntsmwlqtpsluzzsnyyzlmlwvmlnuA
+    │
+    ◆  zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz
+    [EOF]
+    ");
+    let stdout = work_dir
+        .run_jj(&[
+            "run",
+            &format!("'{fake_formatter_path} --uppercase'"),
+            "-r",
+            "..@",
+        ])
+        .success()
+        .stdout;
+    // all commits should be modified
+    insta::assert_snapshot!(stdout, @"");
+}
+
+// This tests a simple `jj run 'cargo fmt' invocation on the repo. It is based
+// on the git-branchless demo here: https://github.com/arxanas/git-branchless/wiki/Command:-git-test
+#[test]
+#[ignore]
+fn test_run_simple_with_cargo() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+    work_dir.write_file(
+        "main.rs",
+        r#"
+    
+    mod foo;
+    
+
+    fn main() {
+        println!("{output"}, output = foo::bar());
+
+    }
+        "#,
+    );
+    work_dir.write_file(
+        "foo.rs",
+        r#"
+      pub fn bar() -> String {
+            "bart".to_owned()
+      }
+
+    "#,
+    );
+
+    work_dir.run_jj(["commit", "-m", "Initial repo"]).success();
+    work_dir
+        .run_jj(["run", "'cargo fmt'", "-r", "root()..@"])
+        .success();
+    let output = work_dir.run_jj(["show", "main.rs"]).success();
+    // main.rs should be nicely formatted now.
+    insta::allow_duplicates! {
+      insta::assert_snapshot!(output.stdout,@r#""#)
+    }
+}
+
+#[test]
+fn test_run_on_immutable() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+    let fake_formatter = assert_cmd::cargo::cargo_bin("fake-formatter");
+    assert!(fake_formatter.is_file());
+    let fake_formatter_path = fake_formatter.to_string_lossy();
+    work_dir.write_file("A.txt", "A");
+    work_dir.run_jj(&["commit", "-m", "A"]).success();
+    work_dir.write_file("b.txt", "b");
+    work_dir.run_jj(&["commit", "-m ", "B"]).success();
+    work_dir.write_file("c.txt", "test to replace");
+    work_dir.run_jj(&["commit", "-m", "C"]).success();
+    insta::assert_snapshot!(get_log_output(&work_dir), @r"
+    @  zsuskulnrvyrovkzqrwmxqlsskqntxvp
+    ○  kkmpptxzrspxrzommnulwmwkkqwworplC
+    │
+    ○  rlvkpnrzqnoowoytxnquwvuryrwnrmlp
+    │
+    ○  qpvuntsmwlqtpsluzzsnyyzlmlwvmlnuA
+    │
+    ◆  zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz
+    [EOF]
+    ");
+    let output = work_dir
+        .run_jj(&[
+            "run",
+            &format!("'{fake_formatter_path} --uppercase'"),
+            "-r",
+            "root()", // Running on the root commit is nonsensical.
+        ])
+        .success();
+    insta::assert_snapshot!(output.stderr, @"");
+    insta::assert_snapshot!(output.stdout, @"");
+}
+
+#[test]
+fn test_run_noop() {
+    let test_env = TestEnvironment::default();
+    test_env.run_jj_in(".", ["git", "init", "repo"]).success();
+    let work_dir = test_env.work_dir("repo");
+    let fake_formatter = assert_cmd::cargo::cargo_bin("fake-formatter");
+    assert!(fake_formatter.is_file());
+    let fake_formatter_path = fake_formatter.to_string_lossy();
+    work_dir.write_file("A.txt", "A");
+    work_dir.run_jj(&["commit", "-m", "A"]).success();
+    work_dir.write_file("b.txt", "b");
+    work_dir.run_jj(&["commit", "-m ", "B"]).success();
+    work_dir.write_file("c.txt", "test to replace");
+    work_dir.run_jj(&["commit", "-m", "C"]).success();
+    insta::assert_snapshot!(get_log_output(&work_dir), @r"
+    @  zsuskulnrvyrovkzqrwmxqlsskqntxvp
+    ○  kkmpptxzrspxrzommnulwmwkkqwworplC
+    │
+    ○  rlvkpnrzqnoowoytxnquwvuryrwnrmlp
+    │
+    ○  qpvuntsmwlqtpsluzzsnyyzlmlwvmlnuA
+    │
+    ◆  zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz
+    [EOF]
+    ");
+    let output = work_dir
+        .run_jj(&[
+            "run",
+            &format!("'{fake_formatter_path} --echo foo --echo $JJ_CHANGE_ID'"),
+            "-r",
+            "..@",
+        ])
+        .success();
+    // As the command does nothing no commits are rewritten.
+    insta::assert_snapshot!(output.stdout, @r"");
+    insta::assert_snapshot!(output.stderr, @r"");
+}
+
+fn get_log_output(work_dir: &TestWorkDir) -> String {
+    work_dir
+        .run_jj(&["log", "-T", r#"change_id ++ description ++ "\n""#])
+        .success()
+        .stdout
+        .to_string()
+}